JS to update user score
Added a function to fetch the answer the user has selected and then clicked on submit. Currently the "increment" variable is global and I need to find a way to make it local to the submit button. Further down the line when it comes to the other questions I need to find a way to reduce mass repetition of code and have a single submit function for all answers.